+++ Makefile
@@ -146,7 +146,7 @@ ifdef DCNN
 	OBJS+=dcnn.o caffe.o
 endif
 # Low-level dependencies last
-SUBDIRS=uct uct/policy playout tactics t-unit t-predict distributed replay patternscan patternplay joseki random montecarlo
+SUBDIRS=uct uct/policy playout tactics t-unit t-predict distributed engines
 
 all: gitversion.h all-recursive pachi
