+++ README
@@ -105,7 +105,7 @@ More information about this dcnn can be found at:
   http://computer-go.org/pipermail/computer-go/2015-December/008324.html
 
 If you want to use a network with different inputs you'll have to tweak
-dcnn.cpp to accomodate it. Pachi will check for "golast19.prototxt" and
+dcnn.c to accomodate it. Pachi will check for "golast19.prototxt" and
 "golast.trained" files on startup and use them if present when
 playing on 19x19. For now dcnn and pondering can't be used together
 (pondering data is thrown away).
