+++ engines/Makefile
@@ -1,0 +1,8 @@
+INCLUDES=-I..
+OBJS=joseki.o josekibase.o montecarlo.o patternplay.o patternscan.o random.o replay.o
+
+all: lib.a
+lib.a: $(OBJS)
+
+
+-include ../Makefile.lib
