+++ engines/patternplay.h
