+++ engines/patternscan.h
