--- montecarlo/Makefile
@@ -1,8 +1,0 @@
-INCLUDES=-I..
-OBJS=montecarlo.o
-
-all: lib.a
-lib.a: $(OBJS)
-
-
--include ../Makefile.lib
