+++ pachi.c
@@ -10,12 +10,12 @@
 #include "board.h"
 #include "debug.h"
 #include "engine.h"
-#include "replay/replay.h"
-#include "montecarlo/montecarlo.h"
-#include "random/random.h"
-#include "patternscan/patternscan.h"
-#include "patternplay/patternplay.h"
-#include "joseki/joseki.h"
+#include "engines/replay.h"
+#include "engines/montecarlo.h"
+#include "engines/random.h"
+#include "engines/patternscan.h"
+#include "engines/patternplay.h"
+#include "engines/joseki.h"
 #include "t-unit/test.h"
 #include "uct/uct.h"
 #include "distributed/distributed.h"
