--- patternplay/Makefile
@@ -1,8 +1,0 @@
-INCLUDES=-I..
-OBJS=patternplay.o
-
-all: lib.a
-lib.a: $(OBJS)
-
-
--include ../Makefile.lib
