--- patternscan/Makefile
@@ -1,8 +1,0 @@
-INCLUDES=-I..
-OBJS=patternscan.o
-
-all: lib.a
-lib.a: $(OBJS)
-
-
--include ../Makefile.lib
